Add Era Summary Key (CMW-531)
